fix "Warning: Unterminated comment starting line" bug
Previously, the method "token_get_all" was used. The problem with
this method is that it generates a warning in cli
which is difficult to fix (one possible approach)
is to install an extension, which is a high cost
for such an unimportant method.

The new approach is far easier, reliable and should
work just fine, since Symfony usually follows PSR4.

See also the comment in the manual:
https://www.php.net/manual/en/function.token-get-all.php#92653